Delete 調整 Controller 路徑/id Service 驗證id存在

+++ TrainingApp/src/main/java/com/Training/Service/PartService.java
@@ -40,6 +40,9 @@ public class PartService {
 
 	// d
 	public void deletePart(String id) {
+		if (!partRepository.existsById(id)) {
+			throw new ConflictException("ID不存在");
+		}
 		partRepository.deleteById(id);
 	}
